🚚 Rename the feature config module
