+++ modules/ppcp-settings/resources/js/Components/Screens/Settings/Components/Overview/features-config.js
